Fix symfony 5.1 deprecations
