+++ .travis.yml
@@ -11,9 +11,9 @@ matrix:
         - php: 7.4
           env: SYMFONY_VERSION=4.4.*
         - php: 7.2
-          env: SYMFONY_VERSION=5.0.*
+          env: SYMFONY_VERSION=5.1.*
         - php: 7.4
-          env: SYMFONY_VERSION=5.0.*
+          env: SYMFONY_VERSION=5.1.*
 
 env:
     global:
